move struct to its own file
